requirements.txt
python-telegram-bot==22.5
google-genai>=1.61.0
Pillow>=12.0.0
python-dotenv>=1.0.0
